Discord buttons

+++ AdventureGameCore/InterfurCreations.AdventureGames.Telegram/DataObjects/Message.cs
@@ -8,7 +8,7 @@ namespace InterfurCreations.AdventureGames.Telegram.DataObjects
 {
     public class Message
     {
-        public int message_id { get; set; }
+        public long message_id { get; set; }
         public From from { get; set; }
         public Chat chat { get; set; }
         public int date { get; set; }
